requirements.txt
django==1.11.7
djangorestframework==3.8.2
